File upload and download

+++ Core/Db/Models/Pizza/ToppingModel.cs
@@ -4,13 +4,13 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public ushort Price { get; set; }
+        public int Price { get; set; }
 
         public ToppingModel()
         {
         }
 
-        public ToppingModel(string name, ushort price)
+        public ToppingModel(string name, int price)
         {
             Name = name;
             Price = price;
