+++ PizzaAPI/Dtos/Pizza/ToppingDto.cs
@@ -3,13 +3,13 @@
     public struct ToppingDto
     {
         public string Name { get; set; }
-        public ushort Price { get; set; }
+        public int Price { get; set; }
 
         public ToppingDto()
         {
         }
 
-        public ToppingDto(string name, ushort price)
+        public ToppingDto(string name, int price)
         {
             Name = name;
             Price = price;
